Removes duplicate declarations of @to_be_foreground and centralizes it un ui.utils
